# Conflicts:
#	src/main/java/com/win/xs_music/controller/UserController.java
#	src/main/java/com/win/xs_music/service/UserService.java
#	src/main/java/com/win/xs_music/service/impl/UserServiceImpl.java

+++ src/main/java/com/win/xs_music/common/BaseContext.java
@@ -6,11 +6,11 @@ package com.win.xs_music.common;
  * @version: 1.0
  */
 public class BaseContext {
-    private static ThreadLocal<Long> threadLocal = new ThreadLocal<>();
-    public static void setCurrentId(Long id) {
+    private static ThreadLocal<Integer> threadLocal = new ThreadLocal<>();
+    public static void setCurrentId(Integer id) {
         threadLocal.set(id);
     }
-    public static Long getCurrentId() {
+    public static Integer getCurrentId() {
         return threadLocal.get();
     }
